This is my first commit into my personal page.
Personal Pages add a bit more background to who you are and
why you are important or relevant. I think I can add a lot of value
to the Latin American scene of Lean Engineering and Agile by focusing
on the following:

-Go a bit more in depth as to why Agile Works. Explain some
great books about the subject, but also talk about the underlying pri-
nciples. It would be great to show reading materials.

-Explain real world examples done in the company about things that we have done
and why/how they have worked.

-Try to explain why things don't work in most places due to rules of power

-Finally, some place to write down ideas for entreprenurship and social advance